requirements.txt
asgiref==3.11.1
Django==4.2.28
djangorestframework==3.16.1
joblib==1.5.3
numpy==2.4.3
psycopg==3.3.3
psycopg-binary==3.3.3
python-dotenv==1.2.2
scikit-learn==1.8.0
scipy==1.17.1
sqlparse==0.5.5
threadpoolctl==3.6.0
